[TWINFACES-313] refactor: remove factory branches screen in a separate folder
